Fix type inferrence for ECDH.get_secret.
This requires explicitly checking for None so that mypy can detect that the
Optional has been resolved and we no longer have uncertain types. This does
somewhat damage the readability of X.can_sign, so the double check is left
in at a slight performance penalty.

I also renamed the "keypair" parameter as I find it more clearly readable
instead of having to realize that self.keypair and keypair are different.

Also avoid the Union type in return by catching and raising on the Inf case.
This may not be strictly correct.